+++ ConcordiumWallet/AppDelegate.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 import Base58Swift
-import MatomoTracker
 import FirebaseMessaging
 import FirebaseCore
 
@@ -64,9 +63,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
                                                object: nil)
         
         UIApplication.shared.statusBarStyle = .lightContent
-        
-        setupMatomoTracker()
-        
+                
         UNUserNotificationCenter.current().delegate = self
         FirebaseApp.configure()
         transactionNotificationService.subscribeToUserDefaultsUpdates()
@@ -174,38 +171,6 @@ extension UIViewController {
 extension UIApplication {
     func topMostViewController() -> UIViewController? {
         return self.keyWindow?.rootViewController?.topMostViewController()
-    }
-}
-
-extension AppDelegate {
-    func setupMatomoTracker() {
-        
-        MatomoTracker.shared.startNewSession()
-        
-        var debug: String {
-            #if DEBUG
-                return "(debug)"
-            #else
-                return ""
-            #endif
-        }
-        
-        var version: String {
-            #if MAINNET
-            if UserDefaults.bool(forKey: "demomode.userdefaultskey".localized) == true {
-                return AppSettings.appVersion + " " + AppSettings.buildNumber + " " + debug
-            }
-            return AppSettings.appVersion
-            #else
-            return AppSettings.appVersion + " " + AppSettings.buildNumber + " " + debug
-            #endif
-        }
-        
-        MatomoTracker.shared.track(view: ["home", "version and network"])
-        
-        MatomoTracker.shared.setDimension(version, forIndex: AppConstants.MatomoTracker.versionCustomDimensionId)
-        MatomoTracker.shared.setDimension(Net.current.rawValue, forIndex: AppConstants.MatomoTracker.networkCustomDimensionId)
-        MatomoTracker.shared.isOptedOut = !UserDefaults.bool(forKey: "isAnalyticsEnabled")
     }
 }