+++ ConcordiumWallet/Features/NewAccountFlow/HomeScreen/HomeScreenView.swift
@@ -111,8 +111,8 @@ struct HomeScreenView: View {
                 Task {
                     await viewModel.reload()
                 }
-                Tracker.track(view: ["Home screen"])
                 updateTimer.start()
+                FirebaseAppTracker.homeScreen()
             }
             .navigationBarTitleDisplayMode(.inline)
             .onDisappear { updateTimer.stop() }
@@ -138,7 +138,6 @@ struct HomeScreenView: View {
                         .onTapGesture {
                             if SettingsHelper.isIdentityConfigured() {
                                 self.router?.showScanQRFlow()
-                                Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Scan QR")
                             } else {
                                 self.router?.showNotConfiguredAccountPopup()
                             }
@@ -229,7 +228,7 @@ struct HomeScreenView: View {
             case .createIdentity:
                 Button(action: {
                     self.router?.showCreateIdentityFlow()
-                    Tracker.trackContentInteraction(name: "Onboarding", interaction: .clicked, piece: "Create Identity")
+                    FirebaseAppTracker.homeIdentityVerificationClicked()
                 }, label: {
                     Text("create_wallet_step_3_title".localized)
                         .font(Font.satoshi(size: 15, weight: .medium))
@@ -244,7 +243,7 @@ struct HomeScreenView: View {
             case .saveSeedPhrase:
                 Button {
                     isShowPasscodeViewShown = true
-                    Tracker.trackContentInteraction(name: "Onboarding", interaction: .clicked, piece: "Save Seed Phrase")
+                    FirebaseAppTracker.homeSaveSeedPhraseClicked()
                 } label: {
                     Text("create_wallet_step_2_title".localized)
                         .font(Font.satoshi(size: 15, weight: .medium))
@@ -368,7 +367,7 @@ struct HomeScreenView: View {
                 self.router?.showNotConfiguredAccountPopup()
             } else {
                 navigationManager.navigate(to: .buy)
-                Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "OnRamp Banner")
+                FirebaseAppTracker.homeOnrampBannerClicked()
             }
         }
         .padding(.horizontal, 16)
@@ -411,7 +410,6 @@ struct HomeScreenView: View {
                 self.router?.showNotConfiguredAccountPopup()
             } else if let selectedAccount = viewModel.selectedAccount?.account as? AccountEntity {
                 navigationManager.navigate(to: .earn(selectedAccount))
-                Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Earn Banner")
             }
         }
         .padding(.horizontal, 16)
@@ -448,30 +446,24 @@ struct HomeScreenView: View {
         let actionItems = [
             ActionItem(iconName: "buy", label: "Buy", action: {
                 navigationManager.navigate(to: .buy)
-                Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Buy")
             }),
             ActionItem(iconName: "send", label: "Send", action: {
                 if let account = viewModel.selectedAccount?.account as? AccountEntity {
                     navigationManager.navigate(to: .send(account, tokenType: .ccd))
-                    Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Send funds")
                 }
             }),
             ActionItem(iconName: "receive", label: "Receive", action: {
                 if let account = viewModel.selectedAccount?.account as? AccountEntity {
                     navigationManager.navigate(to: .receive(account))
-                    Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Account QR")
                 }
             }),
             ActionItem(iconName: "Percent", label: "Earn", action: {
                 guard let selectedAccount = viewModel.selectedAccount?.account as? AccountEntity else { return }
                 navigationManager.navigate(to: .earn(selectedAccount))
-
-                Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Earn")
             }),
             ActionItem(iconName: "activity", label: "Activity", action: {
                 if let account = viewModel.selectedAccount?.account as? AccountEntity {
                     navigationManager.navigate(to: .activity(account))
-                    Tracker.trackContentInteraction(name: "Accounts", interaction: .clicked, piece: "Activity")
                 }
             })
         ]