+++ ConcordiumWallet/Features/NewOnboardingFlow/CreateIdentity/IdentityVerificationView.swift
@@ -50,7 +50,6 @@ final class IdentityVerificationViewModel: ObservableObject {
     }
     
     func selectIdentityProvider(_ ipData: IPInfoResponseElement, pwHash: String) {
-        Tracker.trackContentInteraction(name: "Identity verification", interaction: .clicked, piece: "\(ipData.displayName)")
         self.selectedidentity = ipData
         PermissionHelper.requestAccess(for: .camera) { [weak self] permissionGranted in
             guard let self = self else { return }
@@ -166,6 +165,7 @@ struct IdentityVerificationView: View {
                                 .contentShape(.rect)
                                 .onTapGesture {
                                     self.selectedProvider = provider
+                                    FirebaseAppTracker.identityVerificationScreen(provider: provider.displayName)
                                     withAnimation(.easeInOut(duration: 0.3)) {
                                         isAuthShown.toggle()
                                     }
@@ -209,6 +209,9 @@ struct IdentityVerificationView: View {
                 .padding(.top, 12)
                 .padding(.trailing, 15)
             }
+        }
+        .onAppear {
+            FirebaseAppTracker.identityVerificationProvidersListScreen()
         }
     }