+++ ConcordiumWallet/Features/NewOnboardingFlow/MainPromo/MainPromoView.swift
@@ -41,7 +41,9 @@ struct MainPromoView: View {
         .overlay(alignment: .bottom, content: {
             BottomSheet(isShowing: $isCreateAccountSheetShown) {
                 ActivateAccountSheet()
-                    .onAppear { Tracker.track(view: ["Activate account dialog"]) }
+                    .onAppear {
+                        FirebaseAppTracker.welcomeSetUpWalletDialog()
+                    }
             }
         })
         .fullScreenCover(isPresented: $isCreatePasswordFlowShown) {
@@ -70,7 +72,7 @@ struct MainPromoView: View {
                 .multilineTextAlignment(.center)
             Button {
                 isImportWalletFlowShown.toggle()
-                Tracker.trackContentInteraction(name: "Activate account dialog", interaction: .clicked, piece: "Import Wallet")
+                FirebaseAppTracker.welcomeSetUpWalletDialogImportClicked()
             } label: {
                 VStack(spacing: 2) {
                     Text("import_wallet".localized)
@@ -160,7 +162,7 @@ struct MainPromoView: View {
             }
             Button(action: {
                 isCreatePasswordFlowShown.toggle()
-                Tracker.trackContentInteraction(name: "Activate account dialog", interaction: .clicked, piece: "Create Wallet")
+                FirebaseAppTracker.welcomeSetUpWalletDialogCreateClicked()
             }, label: {
                 HStack {
                     Text("continue_btn_title".localized)